Harden budget trust handling: cache read-only previews, apply immediate trust on tax/austerity changes, dampen neutral   gains, add debug trust/tickbudget tools, and punish over‑taxing with stronger trust loss and income/collection   penalties

+++ src/main/java/dev/citysim/budget/BudgetSnapshot.java
@@ -20,7 +20,7 @@ public class BudgetSnapshot {
     public double toleratedTax;
     public double collectionEfficiency;
     public int trust;
-    public int trustDelta;
+    public double trustDelta;
     public String trustState;
     public boolean austerityEnabled;
